requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
astroid==3.3.11
attrs==25.3.0
bandit==1.8.6
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-option-group==0.5.7
Deprecated==1.2.18
dill==0.4.0
Django==5.2.5
django-environ==0.12.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
dotty-dict==1.3.1
drf-spectacular==0.28.0
flake8==7.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
isort==6.0.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-gitlab==6.3.0
python-semantic-release==10.3.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
smmap==5.0.2
sqlparse==0.5.3
stevedore==5.5.0
tomlkit==0.13.3
types-PyYAML==6.0.12.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wrapt==1.17.3
Faker==13.3.4
